feat: chatgpt docs
