+++ docs/how-to-install-hydraide.md
