+++ docs/install-scripts/docker/docker-compose.swarm.yml
