+++ docs/sdk/go/examples/models/basics_register_swamp.go
@@ -49,6 +49,18 @@ func (m *BasicsRegisterSwamp) RegisterPattern(repo repo.Repo) error {
 	// - Switching from disk-based to in-memory does NOT delete existing SSD data.
 	//   → If you want a clean switch, call `Destroy()` before registering as in-memory.
 
+	// RegisterSwamp always returns a []error.
+	// Each error (if any) represents a failure during Swamp registration on a HydrAIDE server.
+	//
+	// ⚠️ Even when only a single Swamp pattern is registered, HydrAIDE may attempt to replicate or validate
+	// the pattern across multiple server nodes (depending on your cluster).
+	//
+	// ➕ Return behavior:
+	// - If all servers succeeded → returns nil
+	// - If one or more servers failed → returns a non-nil []error
+	//
+	// 🧠 To convert this into a single `error`, you can use the helper:
+	//     hydraidehelper.ConcatErrors(errorResponses)
 	errorResponses := h.RegisterSwamp(ctx, &hydraidego.RegisterSwampRequest{
 
 		// SwampPattern defines which Swamps this config will apply to.
