+++ docs/sdk/go/examples/models/catalog_save_many.go
@@ -93,6 +93,18 @@ func (c *CatalogModelGamePlayer) SaveMany(r repo.Repo, gameID string, players []
 	var newCount, modifiedCount, unchangedCount int
 	var modifiedKeys []string
 
+	// RegisterSwamp always returns a []error.
+	// Each error (if any) represents a failure during Swamp registration on a HydrAIDE server.
+	//
+	// ⚠️ Even when only a single Swamp pattern is registered, HydrAIDE may attempt to replicate or validate
+	// the pattern across multiple server nodes (depending on your cluster).
+	//
+	// ➕ Return behavior:
+	// - If all servers succeeded → returns nil
+	// - If one or more servers failed → returns a non-nil []error
+	//
+	// 🧠 To convert this into a single `error`, you can use the helper:
+	//     hydraidehelper.ConcatErrors(errorResponses)
 	err := h.CatalogSaveMany(ctx, swamp, models, func(key string, status hydraidego.EventStatus) error {
 		switch status {
 		case hydraidego.StatusNew:
