+++ docs/sdk/go/examples/models/catalog_shift_expired.go
@@ -147,7 +147,18 @@ func (m *ModelCatalogQueue) RegisterPattern(repo repo.Repo) error {
 	ctx, cancelFunc := hydraidehelper.CreateHydraContext()
 	defer cancelFunc()
 
-	// Register the Swamp pattern and its configuration
+	// RegisterSwamp always returns a []error.
+	// Each error (if any) represents a failure during Swamp registration on a HydrAIDE server.
+	//
+	// ⚠️ Even when only a single Swamp pattern is registered, HydrAIDE may attempt to replicate or validate
+	// the pattern across multiple server nodes (depending on your cluster).
+	//
+	// ➕ Return behavior:
+	// - If all servers succeeded → returns nil
+	// - If one or more servers failed → returns a non-nil []error
+	//
+	// 🧠 To convert this into a single `error`, you can use the helper:
+	//     hydraidehelper.ConcatErrors(errorResponses)
 	errorResponses := h.RegisterSwamp(ctx, &hydraidego.RegisterSwampRequest{
 		// The pattern applies to all Swamps under the 'queueService/catalog/*' namespace
 		// For example, it matches: queueService/catalog/messages, queueService/catalog/email, etc.
