+++ docs/sdk/go/examples/models/slice_and_reverse_index.go
@@ -335,6 +335,18 @@ func (m *ModelTagProductViewers) RegisterPattern(r repo.Repo, tagName string) er
 	// Retrieve the HydrAIDE SDK instance from the repository.
 	h := r.GetHydraidego()
 
+	// RegisterSwamp always returns a []error.
+	// Each error (if any) represents a failure during Swamp registration on a HydrAIDE server.
+	//
+	// ⚠️ Even when only a single Swamp pattern is registered, HydrAIDE may attempt to replicate or validate
+	// the pattern across multiple server nodes (depending on your cluster).
+	//
+	// ➕ Return behavior:
+	// - If all servers succeeded → returns nil
+	// - If one or more servers failed → returns a non-nil []error
+	//
+	// 🧠 To convert this into a single `error`, you can use the helper:
+	//     hydraidehelper.ConcatErrors(errorResponses)
 	errorResponses := h.RegisterSwamp(ctx, &hydraidego.RegisterSwampRequest{
 
 		// The Swamp pattern name: tags/products/all
